# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Px Lite Production API
  version: 1.0.0
  description: >-
    Px Lite 医药患教内容运营与聚合行为洞察 API。当前契约基于代码现状整理；财务模块仍为前端 demo-only，暂未提供生产 API。
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /content:
    get:
      summary: 获取内容列表
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 分页内容列表
    post:
      summary: 创建内容
      description: 后端能力已存在；前端是否开放完整创建流程待产品确认。
      responses:
        '201':
          description: 内容已创建
        '403':
          description: 无内容写权限
  /content/request-projects:
    get:
      summary: 获取药企选题诉求可关联项目
      description: 药企租户优先返回本租户项目；若 demo 数据未直接绑定租户，则按租户疾病/品牌范围回退返回可演示项目。
      responses:
        '200':
          description: 可关联项目列表
        '403':
          description: 无内容读取权限
  /content/requests:
    get:
      summary: 获取药企选题诉求列表
      description: 支持运营端按状态/项目分页查看 `content_requests`，药企租户仅能查看本租户诉求。
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 诉求分页列表
    post:
      summary: 药企提交选题诉求
      description: 写入 `content_requests`，并生成一条 `requirement_submitted` 草稿内容；不授予药企普通内容写权限。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, requestName, priority, expectedDate, themeFormatMatrix]
              properties:
                projectId:
                  type: string
                requestName:
                  type: string
                priority:
                  type: string
                  enum: [P0, P1, P2]
                expectedDate:
                  type: string
                  format: date
                themeFormatMatrix:
                  type: object
                  additionalProperties:
                    type: object
                    additionalProperties:
                      type: integer
                note:
                  type: string
      responses:
        '201':
          description: 诉求和草稿内容已创建
        '400':
          description: 请求格式无效
        '403':
          description: 无药企诉求提交权限
        '404':
          description: 项目不存在
  /content/requests/{id}:
    get:
      summary: 获取单条选题诉求详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 诉求详情，包含所属项目摘要
        '404':
          description: 诉求不存在
  /content/{id}:
    get:
      summary: 获取内容详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 内容详情
        '404':
          description: 内容不存在
    put:
      summary: 更新内容
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 内容已更新
        '403':
          description: 无内容写权限
        '404':
          description: 内容不存在
    delete:
      summary: 删除内容
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 内容已删除
        '403':
          description: 无内容写权限
...